+++ src/app/controllers/users.ts
@@ -1,23 +1,25 @@
 import express from 'express';
 import { UserMethods } from '../models/user';
 import path from 'path';
+import ReviewModel from '../models/review';
+import sharp from 'sharp';
 // [GET] /users
 export const getAllUsers = async (req: express.Request, res: express.Response): Promise<any> => {
   try {
-    const users = await UserMethods.getUsers();
+    const users = await UserMethods.getUsers().select('+roles');
 
     if (users.length > 0) {
       return res.json({
         status: true,
         message: 'Lấy danh sách người dùng thành công',
-        users,
+        data: users,
       });
     }
 
     return res.json({
       status: true,
       message: 'Danh sách người dùng trống',
-      users,
+      data: users,
     });
   } catch (error) {
     console.log(error);
@@ -30,7 +32,7 @@ export const detailUser = async (req: express.Request, res: express.Response): P
   try {
     const { id } = req.params;
 
-    const user = await UserMethods.getUserById(id);
+    const user = await UserMethods.getUserById(id).select('+roles');
 
     if (!user) {
       return res.status(403).json({
@@ -42,7 +44,7 @@ export const detailUser = async (req: express.Request, res: express.Response): P
     return res.status(200).json({
       status: true,
       message: 'Chi tiết người dùng',
-      user,
+      data: user,
     });
   } catch (error) {
     console.log(error);
@@ -64,7 +66,7 @@ export const searchUsers = async (req: express.Request, res: express.Response): 
       return res.status(200).json({
         status: true,
         message: 'Danh sách người dùng',
-        users,
+        data: users,
       });
     }
 
@@ -82,10 +84,50 @@ export const searchUsers = async (req: express.Request, res: express.Response): 
   }
 };
 
+// [POST] /users
+export const createUser = async (req: express.Request, res: express.Response): Promise<any> => {
+  try {
+    const formData = req.body;
+    const cloneFormData = { ...formData };
+
+    const existingUser = await UserMethods.getUserByEmail(cloneFormData.email);
+
+    if (existingUser && existingUser.email === cloneFormData.email) {
+      return res
+        .json({
+          status: false,
+          message: 'Tài khoản đã tồn tại trong hệ thống, hãy chọn địa chỉ Email khác',
+          existingUser,
+        })
+        .status(403);
+    }
+
+    const user = await UserMethods.createUser(cloneFormData);
+
+    return res.json({
+      status: true,
+      message: 'Thêm mới người dùng thành công',
+      data: user,
+    });
+  } catch (error) {
+    console.log(error);
+    return res.status(500).json({ status: false, message: 'Đã xảy ra lỗi' });
+  }
+};
+
 // [DELETE] /user/:id
 export const deleteUser = async (req: express.Request, res: express.Response): Promise<any> => {
   try {
     const { id } = req.params;
+
+    const existingReviewByUser = await ReviewModel.findOne({ user_id: id });
+
+    if (existingReviewByUser) {
+      return res.json({
+        status: false,
+        message: 'Tồn tại một đánh giá bởi người dùng',
+      });
+    }
 
     await UserMethods.deleteUserById(id);
     const users = await UserMethods.getUsers();
@@ -93,7 +135,7 @@ export const deleteUser = async (req: express.Request, res: express.Response): P
     return res.json({
       status: true,
       message: 'Xoá người dùng thành công',
-      users,
+      data: users,
     });
   } catch (error) {
     console.log(error);
@@ -132,7 +174,7 @@ export const updateUser = async (req: express.Request, res: express.Response): P
 
     await user.save();
 
-    return res.status(200).json({ status: true, message: 'Cập nhật thành công', user });
+    return res.status(200).json({ status: true, message: 'Cập nhật thành công', data: user });
   } catch (error) {
     console.log(error);
     return res.status(500).json({ status: false, message: 'Đã xảy ra lỗi', error });
@@ -141,26 +183,22 @@ export const updateUser = async (req: express.Request, res: express.Response): P
 
 export const uploadAvatar = async (req: express.Request, res: express.Response): Promise<any> => {
   try {
-    const { id } = req.params;
     const imagePath = req.file;
     const CloneImagePath = `/uploads/profile/${imagePath?.originalname}`;
-    const user = await UserMethods.updateUserById(id);
 
-    if (!user) {
-      return res.status(403).json({
-        status: false,
-        message: 'Không tồn tại người dùng cần upload',
+    const fileImage = sharp(CloneImagePath)
+      .resize(262, 317)
+      .toFile(CloneImagePath, function (err: any) {
+        if (err) {
+          console.error('sharp>>>', err);
+        }
+        console.log('ok okoko');
       });
-    }
 
-    user.image = CloneImagePath;
-
-    await user.save();
-
-    return res.status(403).json({
+    return res.status(200).json({
       status: true,
       message: 'Cập nhật ảnh đại diện thành công',
-      user,
+      data: fileImage,
     });
   } catch (error) {
     console.log(error);
